package/libasplib: bump version
libasplib now contains modules, we enable only those for the moment
which do not need external dependencies. We are not in a hurry to fully
support all features of this package because the upcoming Kodi release
17.x-Krypton will not carry any ADSP add-ons:
https://github.com/xbmc/repo-binary-addons/commit/db3b767309df25daae4f01f01dba85ca7b3464b0

-DASPLIB_MODULES_TO_BUILD=some looks hackish because the default is to
enable all modules:
https://github.com/AchimTuran/asplib/blob/master/CMakeLists.txt#L35

But -DASPLIB_MODULES_TO_BUILD=OFF does not disable the build of all
modules as well because of
https://github.com/AchimTuran/asplib/blob/master/asplib_modules/CMakeLists.txt#L19
so I added "some" as fake value.

Signed-off-by: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>