package/coreutils: remove legacy comment
This comment was added in 2007 (e4c6340a: cleanup and fixes), back when
removing a package was (seemingly) supported.

This ability was lost when coreutils was converted to autotarget in 2010
(386183f: coreutils: convert to autotarget), but the comment stuck.

Get rid of it: it nowadays does not make sense and is a bit confusing.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>