configs/snps_archs38_axs103: add u-boot patches to support slave core kick-start
With this patch we add a couple of of-the-tree patches for u-boot.

We have an update in hardware that changed the kick-start slave cores
procedure. So these patches introduce fixes to work with new axs103
version correctly.

They have been already accepted in upstream u-boot so we need to get rid
of them as soon as these updates will get available in buildroot.

Signed-off-by: Vlad Zakharov <vzakhar@synopsys.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>