spidev_test: bump to version in Linux 4.10
spidev_test has further evolved after kernel version 3.15, which
is the most recent version that could be built in this package.
This change allows to build the latest version of spidev_test if
the toolchain's kernel header version allows it. In this way, when
building with a newer toolchain, we are able to benefit from the
added functionality offered by the newest version of the tool.

Signed-off-by: Joeri Barbarien <joeri.barbarien@nokia.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>