configs/pc_x86_64_efi_defconfig: increase ext2 filesystem size
Since commit c6bca8cef0 removed autocalculation of the ext2 filesystem
size, the default size is now set to 60MB. However, this is too small
for pc_x86_64_efi_defconfig. Indeed, the ext2 filesystem contains the
kernel (4MB), the wireless modules (4MB), all firmware for wireless
modules (40MB), and the wifi userspace (9MB) and the udev hwdb (5MB)
which brings the total to 70MB.

Increase the filesystem size to 120000K, which is a nice and round
number and leaves enough space for overhead on a 128MB flash drive.

Fixes:

  https://gitlab.com/buildroot.org/buildroot/builds/15762234

This commit is identical to 9c393ad2fdd2bb6a0955d91bc3d72741b9487b78
from Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>, except it
is done for pc_x86_64_efi_defconfig.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>