configs/armadeus_apf9328_defconfig: remove RPC option
The BR2_TOOLCHAIN_BUILDROOT_INET_RPC has been removed in commit
98f7de8796b282e72d42d3fc4d82d8eec43c6053, following the removal from
upstream uClibc-ng of the RPC support.

However, armadeus_apf9328_defconfig was selecting this option, causing a
build failure due to the selection of a legacy option. We simply remove
it as anyway defconfigs should not needlessly enable toolchain options.

Fixes:

  https://gitlab.com/buildroot.org/buildroot/builds/15762159

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>